# openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoStorm",
    "description": "Observability for AI-driven software discovery",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings/api-key": {
      "post": {
        "tags": [
          "Setup"
        ],
        "summary": "Store Api Key",
        "operationId": "storeApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Storeapikey"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Setup"
        ],
        "summary": "Delete Api Key",
        "operationId": "deleteApiKey",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/setup/autofill": {
      "post": {
        "tags": [
          "Setup"
        ],
        "summary": "Autofill Project",
        "operationId": "autofillProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutofillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutofillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutofillRequest": {
        "properties": {
          "input": {
            "type": "string",
            "minLength": 1,
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "AutofillRequest",
        "description": "Request to autofill project details via AI."
      },
      "AutofillResponse": {
        "properties": {
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "brand_aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Brand Aliases"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "competitors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Competitors"
          },
          "monitoring_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Monitoring Terms"
          }
        },
        "type": "object",
        "required": [
          "project_name",
          "brand_name",
          "description"
        ],
        "title": "AutofillResponse",
        "description": "AI-generated project details."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoreApiKeyRequest": {
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1,
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "StoreApiKeyRequest",
        "description": "Request to store an API key."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}